+++ extensions/keygen/src/interfaces.ts
@@ -1,3 +1,23 @@
+enum TokenKind {
+  Activation = "activation-token",
+  Product = "product-token",
+  User = "user-token",
+  Support = "support-token",
+  Sales = "sales-token",
+  Developer = "developer-token",
+  Admin = "admin-token",
+}
+export interface APIToken {
+  id: string;
+  attributes: {
+    kind: TokenKind;
+    expiry: string | null
+    permissions: string[]
+    name: string;
+    created: string;
+  }
+}
+
 export enum LicenseStatus {
   ACTIVE = "ACTIVE",
   INACTIVE = "INACTIVE",