+++ extensions/keygen/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Keygen Changelog
 
+## [List & Revoke API Tokens] - {PR_MERGE_DATE}
+
+- List API Tokens and Revoke old ones
+
 ## [List Users + API Usage] - 2025-08-12
 
 - List, Create and Delete Users